Trait: GetSpeciesFlagName
Refactors the `GetSpeciesFlagName` trait so it can be used
throughout the whole code (and in the upcoming openPMD ADIOS
PR).
